+++ packages/hpe-design-tokens/src/semantic_color.ts
