+++ packages/hpe-design-tokens/src/semantic_color_normalization.ts
